refactor: DynamicTechBlogComponent 리팩토링
